refactor: types 파일을 따로 빼둠
