--- src/components/PostCard/tempType.ts
@@ -1,7 +1,0 @@
-export interface TempPost {
-  fullName: string;
-  image: string;
-  avatar: string;
-  title: string;
-  count: string;
-}
